config.sh: Create /etc/X11/xorg.conf.d
Since we dropped the Xorg stack from the images, systemd-localed
chokes when trying to set configs because it tries to create
/etc/X11/xorg.conf.d if it doesn't exist so it can write a drop-in
configuration file for setting xkb configuration, and the SELinux policy
does not currently permit it.

Until this is addressed somehow, pre-create the directory so localectl
calls do not fail on this.

Reference: https://bugzilla.redhat.com/2240159