*Desktop: Use the same grub2 config as Fedora anaconda installs
The main part of this is setting `menu_auto_hide=1` in the grubenv file.
This will result in a usually hidden menu with a 1 second timeout. It
enables Fedora's boot success tracking and thus will display the menu
after an unsuccessful boot and will revert to booting the previous
kernel.
This hopefully will result in less support effort after failed kernel
updates as recovery will be more obvious.
Change "GRUB_TIMEOUT" to 5 and omit "GRUB_TIMEOUT_STYLE" in
/etc/default/grub to match Fedora Workstation installs.